Add fields for length, moves and game over.
Add fields to display the current snake length, the number of moves and
whether the game is over.

I'm using the "visibility" style of the "game over" element because
"hidden" means the element isn't displayed but still takes up space. (As
opposed to using the "display" style which collapses the element.)