Add speed powerups
Randomly add blue speed power up "pellets"/"squares". When the snake
"eats" a power up the speed is increased and a timer is created to
remove the increase. Each speed increase has a percentage increase and
an ID and is stored in an array. A timer function is setup to remove the
speed up with the specified ID. Each time a timer is added or removed
the game loop interval is reset with a new interval calculated. This
doesn't work well as the game is no longer smooth.